Dynamic File Change Done

+++ js/all.js
@@ -2,33 +2,27 @@ $(document).ready(function(){
 	
 //動態改標題
 	var check = 0;
-	var a = new Date();
-	var now = a.getFullYear() + '/' + parseInt(a.getMonth()+1) + '/' + a.getDate() + ' ' + a.getHours()+ 
-	':' + a.getMinutes() + ':' +a.getSeconds();
+	
+	var old_value = null;
 
-	$('.topic').on('dblclick', $('.topic'), function(event) {
-		event.preventDefault();
-		if ($(event.target).is('input') || check == 1) {
-			return false;
-		}else{
-			old_value = $(this).html();
-			$(this).html("<input type='text' class='topic_change' value=" + old_value + ">");
-			$('.topic_change[value='+old_value+']').focus();
-			check = 1;
-		}
-	});
+	function getTime(){
+		var a = new Date();
+		var now = a.getFullYear() + '/' + parseInt(a.getMonth()+1) + '/' + a.getDate() + ' ' + a.getHours()+ 
+		':' + a.getMinutes() + ':' +a.getSeconds();
+		return now;
+	}
 
-	$('.topic').on('keyup',$('.topic_change'), function(event) {
-		var id = $(this).data('id');
-		var new_value = $(this).children('.topic_change').val();
+	$('tbody').on('keyup',$('.topic_change'), function(event) {
+		var id = $(event.target).data('id');
 		if (event.keyCode == 13) {
+			var new_value = $(event.target).val();
 			if (old_value==new_value) {
 				alert('你幹嘛要用一樣的名字，87');
 				$('.topic[data-id='+id+']').html(old_value);
 				check = 0;
 				return false;
 			}
-			$.post('./ChangeTopic.php', {id: id, new_value: new_value, now: now}, function(data, textStatus, xhr) {
+			$.post('./ChangeTopic.php', {id: id, new_value: new_value, now: getTime()}, function(data, textStatus, xhr) {
 				if (data == "sametopicname") {
 					alert('標題重複了，87！');
 					$('.topic[data-id='+id+']').html(old_value);
@@ -36,7 +30,7 @@ $(document).ready(function(){
 				}
 				if (data == "success") {
 					$('.topic[data-id='+id+']').html(new_value);
-					$('.time[data-id='+id+']').html(now);
+					$('.time[data-id='+id+']').html(getTime());
 					check = 0;
 				}
 			});
@@ -47,68 +41,184 @@ $(document).ready(function(){
 			check = 0;
 		}
 	});
+
 //動態改標題FIN
 
 //動態改檔案
 //Word File Dynamic Change
 	var old_link = null;
-	$('.doc_file').dblclick(function(event) {
-		var id = $(this).data('id');
+	$('tbody').on('dblclick', function(event) {
+		event.preventDefault();
+		var id = $(event.target).data('id');
+		var target = $(event.target);
 		if ($(event.target).is('input') || check == 1) {
 			return false;
-		}else{
-			old_value = $(this).find('a').attr('href').split('/');
+		}
+
+		if (target.hasClass('topic')) {
+			old_value = $(".topic[data-id="+id+"]").html();
+			$('.topic[data-id='+id+']').html("<input type='text' class='topic_change' data-id="+id+" value='" + old_value + "'>");
+			$('.topic_change[data-id='+id+']').focus();
+			check = 1;
+		}
+
+		if (target.hasClass('doc_file')) {
+			old_value = $(event.target).find('a').attr('href').split('/');
 			old_value = old_value[old_value.length-1];
-			old_link = $(this).html();
-			$(this).html("<input type='file' name='doc_file' class='doc_file_change' data-id="+id+" accept='.doc,.docx'><br>"+"<p>原檔案：" + old_link + "</p>");
+			old_link = $('.doc_file[data-id='+id+']').html();
+			$('.doc_file[data-id='+id+']').html("<input type='file' name='doc_file' class='doc_file_change' data-id="+id+" accept='.doc,.docx'><br>"+"<p>原檔案：" + old_link + "</p>");
 			$('.doc_file_change[data-id='+id+']').click();
+			check = 1;
+		}
+
+		if (target.hasClass('pdf_file')) {
+			old_value = $(event.target).find('a').attr('href').split('/');
+			old_value = old_value[old_value.length-1];
+			old_link = $('.pdf_file[data-id='+id+']').html();
+			$('.pdf_file[data-id='+id+']').html("<input type='file' name='pdf_file' class='pdf_file_change' data-id="+id+" accept='.pdf'><br>"+"<p>原檔案：" + old_link + "</p>");
+			$('.pdf_file_change[data-id='+id+']').click();
+			check = 1;
+		}
+
+		if (target.hasClass('odt_file')) {
+			old_value = $(event.target).find('a').attr('href').split('/');
+			old_value = old_value[old_value.length-1];
+			old_link = $('.odt_file[data-id='+id+']').html();
+			$('.odt_file[data-id='+id+']').html("<input type='file' name='odt_file' class='odt_file_change' data-id="+id+" accept='.odt'><br>"+"<p>原檔案：" + old_link + "</p>");
+			$('.odt_file_change[data-id='+id+']').click();
 			check = 1;
 		}
 	});
 
-	$('.doc_file').on('change',$('.doc_file_change'), function(event) {
+	$('tbody').on('change', function(event) {
 		event.preventDefault();
-		var id = $(this).data('id');
-		var new_value = $(this).children('.doc_file_change').val().split('\\');
+		var target = $(event.target);
+		var id = target.data('id');
+		var new_value = target.val().split('\\');
 		new_value = new_value[new_value.length-1];
-		if (old_value == new_value) {
-			alert('你好像上傳一樣的檔案囉87');
-			$('.doc_file_change').val('');
-			$(this).html(old_link);
-			check = 0;
-			return false;
-		}
-		var data = $('.doc_file_change')[0].files[0];
-		var form_data = new FormData($('.doc_file_change'));
-       	form_data.append("doc_file", data);
-       	form_data.append('id',id);
-       	form_data.append('now',now);
-		$.ajax({
-		    url: 'ChangeDocFile.php',
-		    data: form_data,
-		    cache: false,
-		    contentType: false,
-		    processData: false,
-		    type: 'POST',
-		    success: function(data){
-		    	if (data == "good") {
-		    		$('.doc_file[data-id='+id+']').html("<a href='./upload/" + new_value + "'download>"+ new_value.split('.')[0] +"</a>");
-					$('.doc_download_times[data-id='+id+']').html(0);
-					$('.time[data-id='+id+']').html(now);
-					check = 0;
-		    	}
-		    	if (data == 'notmytype') {
-		    		alert('Only WORD FILE 87');
-		    		$('.doc_file[data-id='+id+']').html(old_link);
-		    		check = 0;
-		    	}
-		    	if (data == 'samefilename') {
-		    		alert('有同檔名的檔案存在了87');
-		    		$('.doc_file[data-id='+id+']').html(old_link);
-		    		check = 0;
-		    	}
+
+		if (target.hasClass('doc_file_change')) {
+			if (old_value == new_value) {
+				alert('你好像上傳一樣的檔案囉87');
+				$('.doc_file_change[data-id='+id+']').val('');
+				$('.doc_file[data-id='+id+']').html(old_link);
+				check = 0;
+				return false;
 			}
-		});
+			var data = $('.doc_file_change')[0].files[0];
+			var form_data = new FormData($('.doc_file_change'));
+	       	form_data.append("doc_file", data);
+	       	form_data.append('id',id);
+	       	form_data.append('now',getTime());
+			$.ajax({
+			    url: 'ChangeDocFile.php',
+			    data: form_data,
+			    cache: false,
+			    contentType: false,
+			    processData: false,
+			    type: 'POST',
+			    success: function(data){
+			    	if (data == "good") {
+			    		$('.doc_file[data-id='+id+']').html("<a href='./upload/" + new_value + "'download>"+ new_value.split('.')[0] +"</a>");
+						$('.doc_download_times[data-id='+id+']').html(0);
+						$('.time[data-id='+id+']').html(getTime());
+						check = 0;
+			    	}
+			    	if (data == 'notmytype') {
+			    		alert('Only WORD FILE 87');
+			    		$('.doc_file[data-id='+id+']').html(old_link);
+			    		check = 0;
+			    	}
+			    	if (data == 'samefilename') {
+			    		alert('有同檔名的檔案存在了87');
+			    		$('.doc_file[data-id='+id+']').html(old_link);
+			    		check = 0;
+			    	}
+				}
+			});
+		}
+
+		if (target.hasClass('pdf_file_change')) {
+			if (old_value == new_value) {
+				alert('你好像上傳一樣的檔案囉87');
+				$('.pdf_file_change').val('');
+				$('.pdf_file[data-id='+id+']').html(old_link);
+				check = 0;
+				return false;
+			}
+			var data = $('.pdf_file_change')[0].files[0];
+			var form_data = new FormData($('.pdf_file_change'));
+	       	form_data.append("pdf_file", data);
+	       	form_data.append('id',id);
+	       	form_data.append('now',getTime());
+			$.ajax({
+			    url: 'ChangePdfFile.php',
+			    data: form_data,
+			    cache: false,
+			    contentType: false,
+			    processData: false,
+			    type: 'POST',
+			    success: function(data){
+			    	if (data == "good") {
+			    		$('.pdf_file[data-id='+id+']').html("<a href='./upload/" + new_value + "'download>"+ new_value.split('.')[0] +"</a>");
+						$('.pdf_download_times[data-id='+id+']').html(0);
+						$('.time[data-id='+id+']').html(getTime());
+						check = 0;
+			    	}
+			    	if (data == 'notmytype') {
+			    		alert('Only PDF FILE 87');
+			    		$('.pdf_file[data-id='+id+']').html(old_link);
+			    		check = 0;
+			    	}
+			    	if (data == 'samefilename') {
+			    		alert('有同檔名的檔案存在了87');
+			    		$('.pdf_file[data-id='+id+']').html(old_link);
+			    		check = 0;
+			    	}
+				}
+			});
+		}
+
+		if (target.hasClass('odt_file_change')) {
+			if (old_value == new_value) {
+				alert('你好像上傳一樣的檔案囉87');
+				$('.odt_file_change').val('');
+				$('.doc_file[data-id='+id+']').html(old_link);
+				check = 0;
+				return false;
+			}
+			var data = $('.odt_file_change')[0].files[0];
+			var form_data = new FormData($('.odt_file_change'));
+	       	form_data.append("odt_file", data);
+	       	form_data.append('id',id);
+	       	form_data.append('now',getTime());
+			$.ajax({
+			    url: 'ChangeOdtFile.php',
+			    data: form_data,
+			    cache: false,
+			    contentType: false,
+			    processData: false,
+			    type: 'POST',
+			    success: function(data){
+			    	if (data == "good") {
+			    		$('.odt_file[data-id='+id+']').html("<a href='./upload/" + new_value + "'download>"+ new_value.split('.')[0] +"</a>");
+						$('.odt_download_times[data-id='+id+']').html(0);
+						$('.time[data-id='+id+']').html(getTime());
+						check = 0;
+			    	}
+			    	if (data == 'notmytype') {
+			    		alert('Only PDF FILE 87');
+			    		$('.odt_file[data-id='+id+']').html(old_link);
+			    		check = 0;
+			    	}
+			    	if (data == 'samefilename') {
+			    		alert('有同檔名的檔案存在了87');
+			    		$('.odt_file[data-id='+id+']').html(old_link);
+			    		check = 0;
+			    	}
+				}
+			});
+		}
 	});
 
 	$('body').on('keyup', $('.doc_file_change'), function(event) {
@@ -120,66 +230,6 @@ $(document).ready(function(){
 		}
 	});
 
-//Pdf File Dynamic Change
-	$('.pdf_file').dblclick(function(event) {
-		var id = $(this).data('id');
-		if ($(event.target).is('input') || check == 1) {
-			return false;
-		}else{
-			old_value = $(this).find('a').attr('href').split('/');
-			old_value = old_value[old_value.length-1];
-			old_link = $(this).html();
-			$(this).html("<input type='file' name='doc_file' class='pdf_file_change' data-id="+id+" accept='.pdf'><br>"+"<p>原檔案：" + old_link + "</p>");
-			$('.pdf_file_change[data-id='+id+']').click();
-			check = 1;
-		}
-	});
-
-	$('.pdf_file').on('change',$('.pdf_file_change'), function(event) {
-		event.preventDefault();
-		var id = $(this).data('id');
-		var new_value = $(this).children('.pdf_file_change').val().split('\\');
-		new_value = new_value[new_value.length-1];
-		if (old_value == new_value) {
-			alert('你好像上傳一樣的檔案囉87');
-			$('.pdf_file_change').val('');
-			$(this).html(old_link);
-			check = 0;
-			return false;
-		}
-		var data = $('.pdf_file_change')[0].files[0];
-		var form_data = new FormData($('.pdf_file_change'));
-       	form_data.append("pdf_file", data);
-       	form_data.append('id',id);
-       	form_data.append('now',now);
-		$.ajax({
-		    url: 'ChangePdfFile.php',
-		    data: form_data,
-		    cache: false,
-		    contentType: false,
-		    processData: false,
-		    type: 'POST',
-		    success: function(data){
-		    	if (data == "good") {
-		    		$('.pdf_file[data-id='+id+']').html("<a href='./upload/" + new_value + "'download>"+ new_value.split('.')[0] +"</a>");
-					$('.pdf_download_times[data-id='+id+']').html(0);
-					$('.time[data-id='+id+']').html(now);
-					check = 0;
-		    	}
-		    	if (data == 'notmytype') {
-		    		alert('Only PDF FILE 87');
-		    		$('.pdf_file[data-id='+id+']').html(old_link);
-		    		check = 0;
-		    	}
-		    	if (data == 'samefilename') {
-		    		alert('有同檔名的檔案存在了87');
-		    		$('.pdf_file[data-id='+id+']').html(old_link);
-		    		check = 0;
-		    	}
-			}
-		});
-	});
-
 	$('body').on('keyup', $('.pdf_file_change'), function(event) {
 		event.preventDefault();
 		var id = $('.pdf_file_change').data('id');
@@ -187,66 +237,6 @@ $(document).ready(function(){
 			$('.pdf_file[data-id='+id+']').html(old_link);
 			check = 0;
 		}
-	});
-
-//Odt File Dynamic Change
-	$('.odt_file').dblclick(function(event) {
-		var id = $(this).data('id');
-		if ($(event.target).is('input') || check == 1) {
-			return false;
-		}else{
-			old_value = $(this).find('a').attr('href').split('/');
-			old_value = old_value[old_value.length-1];
-			old_link = $(this).html();
-			$(this).html("<input type='file' name='odt_file' class='odt_file_change' data-id="+id+" accept='.odt'><br>"+"<p>原檔案：" + old_link + "</p>");
-			$('.odt_file_change[data-id='+id+']').click();
-			check = 1;
-		}
-	});
-
-	$('.odt_file').on('change',$('.odt_file_change'), function(event) {
-		event.preventDefault();
-		var id = $(this).data('id');
-		var new_value = $(this).children('.odt_file_change').val().split('\\');
-		new_value = new_value[new_value.length-1];
-		if (old_value == new_value) {
-			alert('你好像上傳一樣的檔案囉87');
-			$('.odt_file_change').val('');
-			$(this).html(old_link);
-			check = 0;
-			return false;
-		}
-		var data = $('.odt_file_change')[0].files[0];
-		var form_data = new FormData($('.odt_file_change'));
-       	form_data.append("odt_file", data);
-       	form_data.append('id',id);
-       	form_data.append('now',now);
-		$.ajax({
-		    url: 'ChangeOdtFile.php',
-		    data: form_data,
-		    cache: false,
-		    contentType: false,
-		    processData: false,
-		    type: 'POST',
-		    success: function(data){
-		    	if (data == "good") {
-		    		$('.odt_file[data-id='+id+']').html("<a href='./upload/" + new_value + "'download>"+ new_value.split('.')[0] +"</a>");
-					$('.odt_download_times[data-id='+id+']').html(0);
-					$('.time[data-id='+id+']').html(now);
-					check = 0;
-		    	}
-		    	if (data == 'notmytype') {
-		    		alert('Only PDF FILE 87');
-		    		$('.odt_file[data-id='+id+']').html(old_link);
-		    		check = 0;
-		    	}
-		    	if (data == 'samefilename') {
-		    		alert('有同檔名的檔案存在了87');
-		    		$('.odt_file[data-id='+id+']').html(old_link);
-		    		check = 0;
-		    	}
-			}
-		});
 	});
 
 	$('body').on('keyup', $('.odt_file_change'), function(event) {
@@ -258,24 +248,34 @@ $(document).ready(function(){
 		}
 	});
 
-//新增檔案
-	$('.add').click(function(event) {
+//新增Row
+	$('.add').on('click', function(event) {
+		event.preventDefault();
 		var id = null;
-		$.get('./Insert.php?now='+now, function(data) {
+		if (check == 1) {
+	 		return false;
+		}
+
+		$.ajax({
+			url: './Insert.php',
+			type: 'GET',
+			data: {now: getTime()},
+		})
+		.done(function(data) {
 			id = data;
 			$('tbody').append("\
-				<tr>\
+				<tr data-id= " + id + ">\
 				<td data-id= " + id + " class='topic'></td>\
-				<td data-id= " + id + " class='time'>"+ now +"</td>\
-				<td data-id= " + id + " class='doc_file'></td>\
-				<td data-id= " + id + " class='doc_download_times'></td>\
-				<td data-id= " + id + " class='pdf_file'></td>\
-				<td data-id= " + id + " class='pdf_download_times'></td>\
-				<td data-id= " + id + " class='odt_file'></td>\
-				<td data-id= " + id + " class='odt_download_times'></td>\
+				<td data-id= " + id + " class='time'>"+ getTime() +"</td>\
+				<td data-id= " + id + " class='doc_file'><a href=''></a></td>\
+				<td data-id= " + id + " class='doc_download_times'>0</td>\
+				<td data-id= " + id + " class='pdf_file'><a href=''></a></td>\
+				<td data-id= " + id + " class='pdf_download_times'>0</td>\
+				<td data-id= " + id + " class='odt_file'><a href=''></a></td>\
+				<td data-id= " + id + " class='odt_download_times'>0</td>\
 				<td><a href='delete.php?id=$id' class='btn btn-danger'>刪除</a></td>\
 				</tr>");
+			$('.topic_change[data-id='+id+']').focus();
 		});
 	});
-
 })
